lesson6/requirements.txt
aiofiles
alembic
asyncio
asyncpg
greenlet
sqlalchemy
python-dotenv
